lib/proto: a Starlark package for protobuf processing
This change defines an optional Starlark module for encoding
and decoding protocol buffers (https://developers.google.com/protocol-buffers/).

It adds a dependency on google.golang.org/protobuf.

No promises of API stability yet: there are a number of TODOs
to be resolved and questions to be answered first, but I would
like to get this into the hands of the community. See discussion in
https://github.com/stripe/skycfg/issues/23.

Fixes #309

Change-Id: I2f74e266dc2c6b7da34eb0a9414f583105a0852c